requirements.txt
discord.py==2.3.2
python-dotenv==1.0.0
blockfrost-python==0.6.0
aiohttp==3.9.1
chardet==5.2.0
multidict==6.0.4
async-timeout==5.0.1
attrs==23.2.0
yarl==1.9.4
typing-extensions==4.9.0
idna==3.6
psutil==5.9.7
asyncpg==0.29.0
ipaddress==1.0.23
aiodns==3.1.1  # For async DNS resolution
orjson==3.9.10  # Faster JSON parsing
gunicorn==21.2.0  # Required for Heroku
cachetools==5.3.2  # For caching
requests==2.31.0  # For HTTP requests
PyNaCl==1.5.0  # For voice support
certifi==2023.11.17  # For SSL certificate validation
tenacity==8.2.3  # For retry logic
aiofiles==23.2.1  # For async file operations
